Fix bugs identified in review
strcontainsstr: outer loop bound was i < (len-slen), missing the last
valid start position; changed to <=. Inner loop restructured so a match
ending at the final character of the output is detected rather than
falling off the end of the i < len check without returning 1.

clean_salt: parameter and return type changed from char to uint8_t to
avoid implementation-defined signed/unsigned comparison behaviour.

tripcode_2ch: salt buffer changed to uint8_t[] to match; cast to
const char* at the crypt() call.

WAKABA VLA: saltlen and salt are now assigned before the work[] VLA
declaration, so the array is sized with the actual salt length rather
than the initial value of 0.

SHIICHAN read(): check return value and exit on a short read so a
truncated salt file doesn't silently produce wrong tripcodes.

hash.c: SHA-1 uses unsigned throughout but SHA-1 requires exact 32-bit
wrap-around; changed all working variables, block pointer, h[] arrays,
bswap_32, le_bswap_array, and sha1's buffer parameter to uint32_t.
base64: cast uint8_t values to uint32_t before shifting left to avoid
promotion to signed int. sha1: cast length to unsigned before *8 to
avoid signed integer overflow.

crypt.c: cast *key to unsigned char before << 1 to avoid undefined
behaviour when the high bit of a char is set on signed-char platforms.

+++ hash.c
@@ -42,7 +42,7 @@ static void base64(const uint8_t *hash, char *buffer, int length)
 
     for (int i = 0; i < (length/3); i++) {
         int i3 = i*3, i4 = i*4;
-        unsigned bits = hash[i3]<<16 | hash[i3+1]<<8 | hash[i3+2];
+        uint32_t bits = (uint32_t)hash[i3]<<16 | (uint32_t)hash[i3+1]<<8 | hash[i3+2];
 
         buffer[i4+0] = a[(bits>>18) % 64];
         buffer[i4+1] = a[(bits>>12) % 64];
@@ -83,10 +83,10 @@ static void rc4(const uint8_t *input, uint8_t output[6], int length)
 
 #if defined(SHIICHAN)
 static uint32_t rotate(uint32_t i, int n) { return i << n | i >> (32 - n); }
-static void sha1_block(void *input, unsigned h[5])
+static void sha1_block(void *input, uint32_t h[5])
 {
-    unsigned a,b,c,d,e,f,k;
-    unsigned *block = input;
+    uint32_t a,b,c,d,e,f,k;
+    uint32_t *block = input;
     int i;
 
     a = h[0];
@@ -96,7 +96,7 @@ static void sha1_block(void *input, unsigned h[5])
     e = h[4];
 
     for (i = 0; i < 80; i++) {
-        unsigned temp;
+        uint32_t temp;
 
         // FIXME this can be lifted out of loop I think (would cause more memory reads but from L1)
         if (i < 16) {
@@ -138,7 +138,7 @@ static void sha1_block(void *input, unsigned h[5])
     h[4] += e;
 }
 
-static unsigned bswap_32(unsigned n)
+static uint32_t bswap_32(uint32_t n)
 {
 #ifdef __GNUC__
     return __builtin_bswap32(n);
@@ -152,7 +152,7 @@ static unsigned bswap_32(unsigned n)
 static void le_bswap_array(void *a, int words)
 {
 #ifndef __BIG_ENDIAN__
-    unsigned *word = a;
+    uint32_t *word = a;
     while (words--) {
         *word = bswap_32(*word);
         word++;
@@ -161,9 +161,9 @@ static void le_bswap_array(void *a, int words)
 }
 
 // It's assumed that input is allocated a multiple of 64 bytes and we can overwrite it
-static void sha1(uint8_t *input, unsigned *buffer, int length)
+static void sha1(uint8_t *input, uint32_t *buffer, int length)
 {
-    unsigned h[5] = {0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476, 0xC3D2E1F0};
+    uint32_t h[5] = {0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476, 0xC3D2E1F0};
     int tail_length    = 64 - (length % 64);
     int round_length   = length + tail_length;
     uint8_t *input_end = input  + length;
@@ -171,7 +171,7 @@ static void sha1(uint8_t *input, unsigned *buffer, int length)
     bzero(input_end,tail_length);
 
     *input_end = 1 << 7;
-    ((unsigned*)input)[(round_length - 4)/4] = bswap_32(length*8);
+    ((uint32_t*)input)[(round_length - 4)/4] = bswap_32((unsigned)length * 8);
 
     for (int i = 0; i < round_length; i += 64) {
         le_bswap_array(&input[i],16);
